Added to UserController role checking

+++ Presentation/Controllers/UserController.cs
@@ -1,4 +1,6 @@
+using System.Security.Claims;
 using Contracts;
+using Domain.Exceptions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Services.Abstractions;
@@ -17,6 +19,9 @@ public class UsersController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetUsers(CancellationToken cancellationToken)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         var users = await _serviceManager.UserService.GetAllAsync(cancellationToken);
 
         return Ok(users);
@@ -25,6 +30,9 @@ public class UsersController : ControllerBase
     [HttpGet("{userId:guid}")]
     public async Task<IActionResult> GetUserById(Guid userId, CancellationToken cancellationToken)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         var userDto = await _serviceManager.UserService.GetByIdAsync(userId, cancellationToken);
 
         return Ok(userDto);
@@ -33,6 +41,9 @@ public class UsersController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> CreateUser([FromBody] UserForCreationDto userForCreationDto)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         var userDto = await _serviceManager.UserService.CreateAsync(userForCreationDto);
 
         return CreatedAtAction(nameof(GetUserById), new { userId = userDto.Id }, userDto);
@@ -41,6 +52,9 @@ public class UsersController : ControllerBase
     [HttpPut("{userId:guid}")]
     public async Task<IActionResult> UpdateUser(Guid userId, [FromBody] UserForUpdateDto userForUpdateDto, CancellationToken cancellationToken)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         await _serviceManager.UserService.UpdateAsync(userId, userForUpdateDto, cancellationToken);
 
         return NoContent();
@@ -49,6 +63,9 @@ public class UsersController : ControllerBase
     [HttpDelete("{userId:guid}")]
     public async Task<IActionResult> DeleteUser(Guid userId, CancellationToken cancellationToken)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         await _serviceManager.UserService.DeleteAsync(userId, cancellationToken);
 
         return NoContent();
